refactor: split member data into multiple json files
