+++ src/lib/people/team/members/CalebBunye2.json
@@ -1,0 +1,11 @@
+{
+    "name": "Caleb Bunye v2",
+    "title": "President",
+    "committee": "Innovation",
+    "img": "caleb-bunye.webp",
+    "socials": {
+        "github": "https://github.com",
+        "instagram": "https://instagram.com",
+        "website": "https://google.com"
+    }
+}
